+++ libsvmgenerator/src/main/java/nl/tudelft/ir/feature/Feature.java
@@ -6,5 +6,5 @@ import nl.tudelft.ir.index.Document;
 import java.util.List;
 
 public interface Feature {
-    double score(List<String> queryTerms, Document document, Collection collection);
+    float score(List<String> queryTerms, Document document, Collection collection);
 }
